Python3 (#7)
* Added support for python3

The standard recommendation for dealing with bytes in python is "convert to unicode as soon as possible, convert back to bytestream as late as possible". However, in this case I think it's better to keep all data as bytes to make the program oblivious to input and output encodings. The only exception are diagnostic messages, which are assumed to be ascii strings.

* Dropped python2 support, added Logger.char method